+++ cmd/hidden_lake/_template/pkg/config/config_test.go
@@ -1,5 +1,0 @@
-package config
-
-import "testing"
-
-func TestNothing(t *testing.T) {}