+++ cmd/hidden_lake/helpers/template/internal/config/init.go
@@ -3,7 +3,7 @@ package config
 import (
 	"os"
 
-	hl_t_settings "github.com/number571/go-peer/cmd/hidden_lake/_template/pkg/settings"
+	hl_t_settings "github.com/number571/go-peer/cmd/hidden_lake/helpers/template/pkg/settings"
 	logger "github.com/number571/go-peer/internal/logger/std"
 )