+++ cmd/hidden_lake/helpers/template/internal/handler/config_settings_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	hl_t_client "github.com/number571/go-peer/cmd/hidden_lake/_template/pkg/client"
+	hl_t_client "github.com/number571/go-peer/cmd/hidden_lake/helpers/template/pkg/client"
 	testutils "github.com/number571/go-peer/test/_data"
 )